+++ src/custom/cards/sets/SetScripts_Demo2.gd
@@ -65,7 +65,7 @@ func get_scripts(card_name: String) -> Dictionary:
 						"ask_int_max": 5,
 					},
 					{
-						"name": "move_card_cont_to_cont",
+						"name": "move_card_to_container",
 						"src_container": cfc.NMAP.deck,
 						"dest_container": cfc.NMAP.discard,
 						"subject": "index",