+++ tests/integration/test_scripting_engine_costs.gd
@@ -198,7 +198,7 @@ func test_modify_properties_cost():
 
 func test_move_card_cont_to_cont_cost():
 	card.scripts = {"manual": {"hand": [
-			{"name": "move_card_cont_to_cont",
+			{"name": "move_card_to_container",
 			"is_cost": true,
 			"subject": "index",
 			"subject_index": "top",
@@ -215,7 +215,7 @@ func test_move_card_cont_to_cont_cost():
 			"card should turn face-down because "
 			+ "property change cost could be paid")
 	target.scripts = {"manual": {"hand": [
-			{"name": "move_card_cont_to_cont",
+			{"name": "move_card_to_container",
 			"is_cost": true,
 			"subject": "index",
 			"subject_index": "top",
@@ -234,7 +234,7 @@ func test_move_card_cont_to_cont_cost():
 
 func test_move_card_cont_to_board_cost():
 	card.scripts = {"manual": {"hand": [
-			{"name": "move_card_cont_to_board",
+			{"name": "move_card_to_board",
 			"is_cost": true,
 			"subject": "index",
 			"subject_index": "top",
@@ -249,9 +249,9 @@ func test_move_card_cont_to_board_cost():
 	yield(yield_to(card._flip_tween, "tween_all_completed", 0.4), YIELD)
 	assert_false(card.is_faceup,
 			"card should turn face-down because "
-			+ "property change cost could be paid")
+			+ "enough cards could be moved from the deck")
 	target.scripts = {"manual": {"hand": [
-			{"name": "move_card_cont_to_board",
+			{"name": "move_card_to_board",
 			"is_cost": true,
 			"subject": "index",
 			"subject_index": "top",
@@ -266,4 +266,41 @@ func test_move_card_cont_to_board_cost():
 	yield(yield_to(target._flip_tween, "tween_all_completed", 0.4), YIELD)
 	assert_true(target.is_faceup,
 			"card should stay face-up because "
-			+ "property change cost could not be paid")
+			+ "not enough cards could be found in the deck")
+
+func test_move_card_to_grid_cost():
+	card.scripts = {"manual": {"hand": [
+			{"name": "move_card_to_board",
+			"is_cost": true,
+			"subject": "index",
+			"subject_index": "top",
+			"subject_count": 3,
+			"src_container":  cfc.NMAP.deck,
+			"grid_name":  "BoardPlacementGrid"},
+			{"name": "flip_card",
+			"subject": "self",
+			"set_faceup": false}]}}
+	card.execute_scripts()
+	yield(yield_to(card._flip_tween, "tween_all_completed", 0.4), YIELD)
+	yield(yield_to(card._flip_tween, "tween_all_completed", 0.4), YIELD)
+	assert_false(card.is_faceup,
+			"card should turn face-down because "
+			+ "grid had enough slots for all cards")
+	target._debugger_hook = true
+	target.scripts = {"manual": {"hand": [
+			{"name": "move_card_to_board",
+			"is_cost": true,
+			"subject": "index",
+			"subject_index": "top",
+			"subject_count": 3,
+			"src_container":  cfc.NMAP.deck,
+			"grid_name":  "BoardPlacementGrid"},
+			{"name": "flip_card",
+			"subject": "self",
+			"set_faceup": false}]}}
+	target.execute_scripts()
+	yield(yield_to(target._flip_tween, "tween_all_completed", 0.4), YIELD)
+	yield(yield_to(target._flip_tween, "tween_all_completed", 0.4), YIELD)
+	assert_true(target.is_faceup,
+			"card should stay face-up because "
+			+ "grid did not have enough slots for all cards")